+++ app/mailers/account_mailer.rb
@@ -12,10 +12,26 @@ class AccountMailer < ApplicationMailer
     mail(to: ENV["ADMIN_EMAIL"], subject: "New sign up at Abroaders app - #{@account.email}")
   end
 
+  # `timestamp` is an Unix integer timestamp, not a Date object, because the
+  # latter can't be stored in Redis
   def notify_admin_of_survey_completion(account_id, timestamp)
     @account = Account.find(account_id)
     @timestamp = Time.at(timestamp).in_time_zone("EST")
     mail(to: ENV["ADMIN_EMAIL"], subject: "App Profile Complete - #{@account.email}")
+  end
+
+  # `timestamp` is an Unix integer timestamp, not a Date object, because the
+  # latter can't be stored in Redis.
+  #
+  # This is triggered when a person who is NOT ready updates their status to
+  # ready. It's NOT triggered when a new person says on the onboarding survey
+  # that they are ready.
+  def notify_admin_of_user_readiness_update(account_id, timestamp)
+    @account   = Account.find(account_id)
+    @owner     = @account.owner
+    @companion = @account.companion
+    @timestamp = Time.at(timestamp).in_time_zone("EST")
+    mail(to: ENV["ADMIN_EMAIL"], subject: "User is Ready - #{@account.email}")
   end
 
 end